Rework JWT-Test-Cases to work without devon4j-kafka
