+++ starters/starter-security-jwt/src/test/java/com/devonfw/module/security/jwt/common/base/kafka/KafkaMessageListener.java
@@ -1,15 +1,13 @@
 package com.devonfw.module.security.jwt.common.base.kafka;
 
-import javax.inject.Inject;
+import java.util.ArrayList;
+import java.util.List;
+
 import javax.inject.Named;
 
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
-
-import com.devonfw.module.kafka.common.messaging.retry.api.client.MessageProcessor;
-import com.devonfw.module.kafka.common.messaging.retry.api.client.MessageRetryOperations;
-import com.devonfw.module.security.jwt.common.base.kafka.JwtAuthentication;
 
 /**
  * A Listener class with {@link KafkaListener} listens the message for the given topic and group name.
@@ -18,11 +16,23 @@ import com.devonfw.module.security.jwt.common.base.kafka.JwtAuthentication;
 @Named
 public class KafkaMessageListener {
 
-  @Inject
-  private MessageRetryOperations<String, String> messageRetryOperations;
+  List<ConsumerRecord<String, String>> receivedMessages = new ArrayList<>();
 
-  @Inject
-  private MessageProcessor<String, String> messageProcessor;
+  /**
+   * @return receivedMessages
+   */
+  public List<ConsumerRecord<String, String>> getReceivedMessages() {
+
+    return this.receivedMessages;
+  }
+
+  /**
+   * @param receivedMessages new value of {@link #getReceivedMessages}.
+   */
+  public void setReceivedMessages(List<ConsumerRecord<String, String>> receivedMessages) {
+
+    this.receivedMessages = receivedMessages;
+  }
 
   /**
    * This method is used to listen the message in kafka broker for the given topic and group name in
@@ -34,18 +44,9 @@ public class KafkaMessageListener {
    */
   @JwtAuthentication(failOnMissingToken = true)
   @KafkaListener(topics = JwtTokenValidationAspectTest.TEST_TOPIC, groupId = "test-group")
-  public void consumer(ConsumerRecord<String, String> consumerRecord, Acknowledgment acknowledgment) throws Exception {
+  public void consume(ConsumerRecord<String, String> consumerRecord, Acknowledgment acknowledgment) throws Exception {
 
-    processMessageAndAcknowledgeListener(consumerRecord, acknowledgment);
-  }
-
-  private void processMessageAndAcknowledgeListener(ConsumerRecord<String, String> consumerRecord,
-      Acknowledgment acknowledgment) {
-
-    this.messageRetryOperations.processMessageWithRetry(consumerRecord, this.messageProcessor);
-
-    // Acknowledge the listener.
-    acknowledgment.acknowledge();
+    this.receivedMessages.add(consumerRecord);
   }
 
   /**
@@ -57,10 +58,9 @@ public class KafkaMessageListener {
    * @throws Exception exception
    */
   @JwtAuthentication(failOnMissingToken = false)
-  @KafkaListener(topics = JwtTokenValidationAspectTest.TEST_TOPIC_2, groupId = "test-group-2")
-  public void consumer2(ConsumerRecord<String, String> consumerRecord, Acknowledgment acknowledgment) throws Exception {
+  @KafkaListener(topics = JwtTokenValidationAspectTest.TEST_TOPIC_2, groupId = "test-group")
+  public void consume2(ConsumerRecord<String, String> consumerRecord, Acknowledgment acknowledgment) throws Exception {
 
-    processMessageAndAcknowledgeListener(consumerRecord, acknowledgment);
+    this.receivedMessages.add(consumerRecord);
   }
-
 }
